# Conflicts:
#	Assets/Library/SaveLoad/SaveLoadManager.cs
#	Assets/Test/2ENO/DunGeonMap/DungeonSystem.cs
#	Assets/Test/2ENO/DunGeonMap/SaveData/DungeonData.cs
#	Assets/Test/AS/SwipeSomething/SwipeSomething.cs
#	Assets/Test/TestInventory/DataScript/UserData.cs

+++ Assets/Test/2ENO/ConsumeManager/SaveData/ConsumableSaveData_0.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ConsumableSaveData_0 : SaveDataBase
 {
-    public int curStamina;
+    public float curStamina;
     public int curIngameHour;
     public int curIngameMinute;
     public TimeState curTimeState;
